01-05 20:30 천석희 일일 운동량 조회에서 루틴으로 등록할 수 있게 모달구현해야함

+++ fineapple/src/main/java/kr/or/fineapple/domain/Exer.java
@@ -20,7 +20,7 @@ public class Exer {
 	String exerVideoTime;
 	String dodbogi;
 	String exerLink;
-	Double recommTime;
+	int recommTime;
 	String anExerTime;
